+++ components/google-cloud/google_cloud_pipeline_components/experimental/hyperparameter_tuning_job/__init__.py
@@ -1,4 +1,4 @@
-# Copyright 2021 The Kubeflow Authors. All Rights Reserved.
+# Copyright 2023 The Kubeflow Authors. All Rights Reserved.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -13,10 +13,21 @@
 # limitations under the License.
 """Module for supporting Google Vertex AI Hyperparameter Tuning Job Op."""
 
-import os
 
-from .utils import serialize_parameters, serialize_metrics, GetTrialsOp, GetBestTrialOp, GetBestHyperparametersOp, GetHyperparametersOp, GetWorkerPoolSpecsOp, IsMetricBeyondThresholdOp
-from kfp.components import load_component_from_file
+from google_cloud_pipeline_components.v1.hyperparameter_tuning_job import (
+    HyperparameterTuningJobRunOp,
+    serialize_metrics,
+    serialize_parameters,
+)
+
+from .utils import (
+    GetBestHyperparametersOp,
+    GetBestTrialOp,
+    GetHyperparametersOp,
+    GetTrialsOp,
+    GetWorkerPoolSpecsOp,
+    IsMetricBeyondThresholdOp,
+)
 
 __all__ = [
     'HyperparameterTuningJobRunOp',
@@ -29,8 +40,3 @@ __all__ = [
     'serialize_parameters',
     'serialize_metrics',
 ]
-
-
-HyperparameterTuningJobRunOp = load_component_from_file(
-    os.path.join(os.path.dirname(__file__), 'component.yaml')
-)
