+++ components/google-cloud/google_cloud_pipeline_components/experimental/hyperparameter_tuning_job/utils.py
@@ -14,7 +14,6 @@
 # fmt: off
 """Module for supporting Google Vertex AI Hyperparameter Tuning Job Op."""
 
-from google.cloud.aiplatform_v1.types import study
 from kfp import dsl
 
 
@@ -35,9 +34,11 @@ def GetTrialsOp(gcp_resources: str) -> list:
       List of strings representing the intermediate JSON representation of the
       trials from the hyperparameter tuning job.
   """
-  from google.cloud import aiplatform
-  from google_cloud_pipeline_components.proto.gcp_resources_pb2 import GcpResources
+  from google_cloud_pipeline_components.proto.gcp_resources_pb2 import \
+      GcpResources
+
   from google.protobuf.json_format import Parse
+  from google.cloud import aiplatform
   from google.cloud.aiplatform_v1.types import study
 
   api_endpoint_suffix = '-aiplatform.googleapis.com'
@@ -242,57 +243,3 @@ def GetWorkerPoolSpecsOp(best_hyperparameters: list,
           f'--{p.parameter_id}={p.value}')
 
   return worker_pool_specs
-
-
-def serialize_parameters(parameters: dict) -> list:
-  """Serializes the hyperparameter tuning parameter spec to dictionary format.
-
-  Args:
-      parameters (Dict[str, hyperparameter_tuning._ParameterSpec]):
-        Dictionary representing parameters to optimize. The dictionary key is the
-        parameter_id, which is passed into your training job as a command line
-        key word argument, and the dictionary value is the parameter
-        specification of the metric. from google.cloud.aiplatform
-        import hyperparameter_tuning as hpt
-        parameters={
-            'decay': hpt.DoubleParameterSpec(min=1e-7, max=1, scale='linear'),
-            'learning_rate': hpt.DoubleParameterSpec(min=1e-7, max=1, scale='linear')
-            'batch_size': hpt.DiscreteParamterSpec(values=[4, 8, 16, 32, 64, 128], scale='linear') }
-        Supported parameter specifications can be found in aiplatform.hyperparameter_tuning.
-        These parameter specification are currently supported: DoubleParameterSpec, IntegerParameterSpec, CategoricalParameterSpace, DiscreteParameterSpec
-        Note: The to_dict function is used here instead of the to_json
-        function for compatibility with GAPIC.
-
-  Returns:
-      List containing an intermediate JSON representation of the parameter spec
-
-  """
-  return [
-      study.StudySpec.ParameterSpec.to_dict(
-          parameter._to_parameter_spec(parameter_id=parameter_id))
-      for parameter_id, parameter in parameters.items()
-  ]
-
-
-def serialize_metrics(metric_spec: dict) -> list:
-  """Serializes a metric spec to dictionary format.
-
-  Args:
-      metric_spec (Dict[str, str]):
-        Required. Dictionary representing metrics
-        to optimize. The dictionary key is the metric_id, which is reported by
-        your training job, and the dictionary value is the optimization goal of
-        the metric ('minimize' or 'maximize'). Example:
-        metrics = {'loss': 'minimize', 'accuracy': 'maximize'}
-
-  Returns:
-      List containing an intermediate JSON representation of the metric spec
-
-  """
-  return [
-      study.StudySpec.MetricSpec.to_dict(
-          study.StudySpec.MetricSpec({
-              'metric_id': metric_id,
-              'goal': goal.upper()
-          })) for metric_id, goal in metric_spec.items()
-  ]
