+++ components/google-cloud/tests/experimental/hyperparameter_tuning_job/integration/test_util_components_compile.py
@@ -16,11 +16,19 @@
 import json
 import os
 import re
-from typing import Dict, Any
-import yaml
+from typing import Any, Dict
 
-from google_cloud_pipeline_components.experimental.hyperparameter_tuning_job import HyperparameterTuningJobRunOp, serialize_metrics, GetTrialsOp, GetBestTrialOp, GetBestHyperparametersOp, GetHyperparametersOp, GetWorkerPoolSpecsOp, IsMetricBeyondThresholdOp
 import kfp
+import yaml
+from google_cloud_pipeline_components.experimental.hyperparameter_tuning_job import (
+    GetBestHyperparametersOp,
+    GetBestTrialOp,
+    GetHyperparametersOp,
+    GetTrialsOp,
+    GetWorkerPoolSpecsOp,
+    IsMetricBeyondThresholdOp,
+    serialize_metrics,
+)
 from kfp import compiler
 
 import unittest
